refactor: move generateVars into dedicated file
